looking-glass-inputs/requirements.txt
numpy
matplotlib
watchdog
flask
web3
python-dotenv
